+++ src/metrics/http.rs
@@ -4,7 +4,6 @@ use std::{
     fmt::Display,
     future::Future,
     pin::Pin,
-    str::FromStr,
     sync::Arc,
     task::{ready, Context, Poll},
     time::Instant,
@@ -260,13 +259,11 @@ impl ResponseMetricState {
             match side {
                 // For client side the protocol is the URL.
                 MetricSide::Client => {
-                    let url_scheme = req.uri().scheme_str();
-                    let server_address = req.uri().host();
-                    let server_port = req.uri().port_u16().or(match url_scheme {
-                        Some("http") => Some(util::HTTP_DEFAULT_PORT),
-                        Some("https") => Some(util::HTTPS_DEFAULT_PORT),
-                        _ => None,
-                    });
+                    let util::HttpRequestAttributes {
+                        url_scheme,
+                        server_address,
+                        server_port,
+                    } = util::HttpRequestAttributes::from_sent_request(req);
 
                     if let Some(server_address) = server_address {
                         attributes
@@ -280,42 +277,11 @@ impl ResponseMetricState {
                     }
                 }
                 MetricSide::Server => {
-                    let (host, url_scheme) = req
-                        .headers()
-                        .get(http::header::FORWARDED)
-                        .map(util::Forwarded::parse_header_value)
-                        .map(|util::Forwarded { host, proto }| (host, proto))
-                        .unwrap_or_default();
-
-                    let host = host
-                        .or_else(|| {
-                            req.headers()
-                                .get(util::X_FORWARDED_HOST)
-                                .and_then(|v| v.to_str().ok())
-                        })
-                        .or_else(|| {
-                            req.headers()
-                                .get(http::header::HOST)
-                                .and_then(|v| v.to_str().ok())
-                        });
-
-                    let url_scheme = url_scheme.or_else(|| {
-                        req.headers()
-                            .get(util::X_FORWARDED_PROTO)
-                            .and_then(|v| v.to_str().ok())
-                    });
-
-                    let (server_address, server_port) = host
-                        .and_then(|host| host.split_once(':'))
-                        .map_or_else(|| (host, None), |(host, port)| (Some(host), Some(port)));
-
-                    let server_port = server_port
-                        .and_then(|server_port| u16::from_str(server_port).ok())
-                        .or(match url_scheme {
-                            Some("http") => Some(util::HTTP_DEFAULT_PORT),
-                            Some("https") => Some(util::HTTPS_DEFAULT_PORT),
-                            _ => None,
-                        });
+                    let util::HttpRequestAttributes {
+                        url_scheme,
+                        server_address,
+                        server_port,
+                    } = util::HttpRequestAttributes::from_recv_request(req);
 
                     if let Some(server_address) = server_address {
                         attributes