+++ CHANGELOG.md
@@ -4,9 +4,9 @@
   value is extracted from the request URL. For server request the value is
   extracted from `X-Forwarded-Proto` and `Forwarded` headers.
 
-- Add `server.address` and `server.port` attributes to HTTP spans. For client
-  request the value is extracted from the request URL. For server request the
-  value is extracted from request following [these rules].
+- Add `server.address` and `server.port` attributes to HTTP spans/metrics and
+  gRPC spans. For client request the value is extracted from the request URL.
+  For server request the value is extracted from request following [these rules].
 
 ## v0.6.2